Set up C tests to fail upon warnings (#14937)
* Fix string formatting and inaccurate types in C tests

In a couple of places we were using wrong string formatters (e.g. %d instead of %zu), sometimes the number of formatters didn't match the arguments, and sometimes the inline type annotation was too broad (int[] instead of uint32_t[]).

* Enable `-Wall -Wextra -Werror` to fail eagerly

... this did uncover some bugs

* extra sausage for windows

* baptize the newcomers

* enable -Wshadow for MSVC consistency

sounds like a good thing to enable anyways

* downgrade MSVC to W3

W4 includes a bunch of pedantic warnings that are not included in GCCs -Wall -Wextra

* downgrade MSVC to W2

* missing void, more wrong formatters

* on second thought: remove MSVC only check on GCC

* note on clang

+++ test/c/test_split_2q_unitaries.c
@@ -41,7 +41,7 @@ int test_split_2q_unitaries_no_unitaries(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "cx");
         if (gate != 0) {
             printf("gates changed when there should be no circuit changes\n");
@@ -81,14 +81,14 @@ int test_split_2q_unitaries_x_y_unitary(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "unitary");
         if (gate != 0) {
             printf("Gates outside expected set in output circuit\n");
             result = EqualityError;
             goto ops_cleanup;
         }
-        unsigned int count = counts.data[i].count;
+        size_t count = counts.data[i].count;
         if (count != 2) {
             printf("Unexpected gate counts found\n");
             result = EqualityError;
@@ -148,14 +148,14 @@ int test_split_2q_unitaries_swap_x_y_unitary(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "unitary");
         if (gate != 0) {
             printf("Gates outside expected set in output circuit\n");
             result = EqualityError;
             goto ops_cleanup;
         }
-        unsigned int count = counts.data[i].count;
+        size_t count = counts.data[i].count;
         if (count != 2) {
             printf("Unexpected gate counts found\n");
             result = EqualityError;
